fix website build failure due to suppress_comment
Summary:
the website build runs flow from master to generate the inline examples. in master, `suppress_comment` is no longer a valid option, causing the build to fail.

`$DocIssue` was a suppession we used in the past when we had expected errors that we didn't want to show in the docs, like when stubbing out parts of the examples.

Reviewed By: dsainati1

Differential Revision: D21859589

fbshipit-source-id: 524e5f7af212f0b990bc92fefa7561886e414785