Move instantiation cache global into context
Summary:
This is one small piece of global state which is now slightly more local.

This no longer needs to be cleared as part of Flow_cache.clear, at least for
correctness purposes. This cache will not interfere with another context, and
will be collected when the host context is.

Reviewed By: dsainati1

Differential Revision: D21633543

fbshipit-source-id: cc9e2ac51f9eb347430c41be68b3125acafde460