Move repos cache global into context
Summary:
This is one small piece of global state which is now slightly more local.

This no longer needs to be cleared as part of Flow_cache.clear, at least for
correctness purposes. This cache will not interfere with another context, and
will be collected when the host context is.

Reviewed By: dsainati1

Differential Revision: D21633545

fbshipit-source-id: 342464a02e30608e98da9c7d0b591b51f1cd9649