Move constraint cache global state into context
Summary:
This is one small piece of global state which is now slightly more local.

This no longer needs to be cleared as part of Flow_cache.clear, at least for
correctness purposes. This cache will not interfere with another context, and
will be collected when the host context is.

Reviewed By: dsainati1

Differential Revision: D21633542

fbshipit-source-id: a5045a87e0ad92174e35d914b184e8e7270b5a35