Give an annot reason to Array<> annotations
Summary:
Type annotations should carry around an "annot" location, which is insensitive to
repositioning. In various places, this location is used to provide better error
messages. The differentiating information is also important if the reason is
part of a cache key.

The small regression in React createClass errors is addressed in a follow-up.

Reviewed By: dsainati1

Differential Revision: D21754470

fbshipit-source-id: 029b237625bf1b8941a31ca883a6f588908d3cba