Tweak creation of reasons for typeapps
Summary:
There are a few different reasons we might want to create a TypeAppT type, which
was reflected in the API of TypeUtil.typeapp, with its optional arguments and
various conditional logic.

This diff ~~foolishly~~boldly attempts to change two things at once:
1. Split up the API into separate function calls for each distinct use
2. Improve the reasons created for "builtin typeapps"

Consider the ArrT type, which represents a type application of the Array<T>
class. We model this as a more specialized type, but in other places we want to
rely on the libdef -- method calls, for example. In those cases, we convert the
ArrT to a TypeAppT of the builtin class type.

In this case, it's desirable to keep most of the reason information associated
with the original ArrT, including both the assoiated annot loc, if applicable,
and the repositioned loc.

The other, more common case, is type annotations. In this case, we definitely
want an annot loc, but the current/repos loc is the same loc.

Lastly, we have implicit typeapps, which I've left alone, since the expected
logic seems to be implementation defined.

Reviewed By: panagosg7

Differential Revision: D21794195

fbshipit-source-id: 32b4ac30801e5f26ae7de7b6a277685fca7a6c91