Move speculation state global into context
Summary:
This is a larger change than previous work in this space, but the motivation is
the same.

The Speculation module already depends on Context, so it is not possible to
simply add add the speculation state to the context, since that would introduce
a trivially cyclic dependency.

This diff splits the Speculation module into two separate modules:

1. Speculation_state, which includes only the type definitions necessary to
   describe the current state of speculation.
2. Speculation, the remainder of the original module, which includes logic to
   find the relevant tvars of an action, diff between cases, etc.

I've made a few small changes as a part of this diff:

* I removed the Action, Case, State modules and replaced with more naive
  namespacing via prefixes. This reduces the need to locally open modules, which
  ended up being pretty cumbersome in practice.
* I added a speculation.mli file, to document and hide the implementation details of the
  Speculation module.
* I tidied up some of the callers into this code from Flow_js. It seems like
  ocamlformat made a bit of a mess here.

Reviewed By: dsainati1

Differential Revision: D21633539

fbshipit-source-id: 2bf715adad7fc6493fd9749b72acf39a16845b6e