Treat createClass component propTypes as annots
Summary:
Flow has support for defining components using createClass, which is long
deprecated but still has some uses in our codebase. Using createClass,
programmers declare the expected props of their component by defining an object
value, which Flow converts into an annotation.

The exact method for this conversion is not relevant here. The important part is
that we have values which should be treated more like annotations. This diff
ensures that the reasons for these types have an associated "annot location."
This results in better error messages, as can be seen in the changed test
output.

Reviewed By: jbrown215

Differential Revision: D21754471

fbshipit-source-id: 2fbc40a982963facd74d1a2cf5e33d65c31a36b8