[Errors] parse codes from suppression comments
Summary:
Extend regex for checking comments to extract codes from suppression comments.

Each suppression can have 1 or 0 codes. Codes must be of the form `[code-name]`. If a suppression does not contain a code, we treat it as suppressing all errors, while suppressions that contain a code only suppress errors of that code-type.

Reviewed By: pieterv

Differential Revision: D21164303

fbshipit-source-id: 41b695a74517ff675fa6da12957697d683ff6a56