[Errors] Check for code overlap when determining if an error is suppressed
Summary:
This diff does two things:

First, we collect all the suppression comments in a file and assign them to the locations they suppress. Since each comment can contain at most one code, we combine stacked suppressions into a set of codes that apply to the first non-suppression line of code. These are collected in a set and added to the suppression map.

Second, when determining whether an error is suppressed, we look up all the codes that apply in a given suppression and subtract out the relevant codes for the error in question. Note that because multiple errors can occur on a line, we may require multiple suppressions to suppress a single line.

Suppressions without codes still suppress any error.

Reviewed By: pieterv

Differential Revision: D19760726

fbshipit-source-id: 296ed9d77d310a5056d75e0a5c0e4ded642810c0