Lint rule for renaming export to have name `default`
Summary:
There is one more lint rule we would like to enforce for the new ES6 imports/exports for consistency with other rules around the default export, to avoid treating the default export as just a named export with name "default". This last rule is that the default export of a module must be specified with `export default Foo` syntax, instead of another value being renamed to the default export using `export {Foo as default}`.

```
// Allowed
export default Foo;

// Not allowed
export {Foo as default}
```

This new lint rule is called `export-renamed-default` and is gated behind the `experimental.strict_es6_import_export` flowconfig flat to assist with testing and rollout.

To implement this we simply need to check for any renamed exports that have been renamed to default, and then error appropriately.

Reviewed By: pieterv

Differential Revision: D21895068

fbshipit-source-id: 69cfc4ce4271a314fd163d2ea67b06c1be6641a5