update add-comments to support error codes
Summary:
Update the `add-comments` script to include error codes in its added suppressions.

Note that this changes the API of the tool. Previously, when specifying a suppression, it was necessary to include the suppressor in the comment text when running the command, e.g:
 `./tool add-comments --comment="\$FlowFixMe etc etc` would insert a comment of the form `//$FlowFixMe etc etc`

Now that suppression syntax is standardized, and in order to allow for insertion of a code after the suppressor, one only needs to specify the post-suppression text of the comment. For example:

 `./tool add-comments --comment="etc etc` would insert a comment of the form `//$FlowFixMe[CODE] etc etc` for each error code that appears on a line. This means that potentially invoking the comment could insert multiple suppression comments on a single line of code.

Reviewed By: mroch

Differential Revision: D21847076

fbshipit-source-id: bd60ead1b10c97c51f857379af15143aee4b0350