Enable LSP support for autofix exports by default
Summary:
This enabled the "autofix export" code action by default. Before, this was
available under a little-known option (`experimental.autofix_exports=true`).

With this change, this action will be available whenever "well_formed_exports"
is enabled for a particular file. This means that it respects the whitelist option
if present.

The `experimental.autofix_exports` is deleted.

Reviewed By: nmote

Differential Revision: D21915131

fbshipit-source-id: 9d1290219e5baa8e230a4685889a3f3ba3069ce6