[autofix] Reuse "annotate-exports" import type machinery
Summary:
I have ported the annotate exports imports utilities to insert-type, and
added a switch that forces an import to occur even when it will cause
a name collision or isn't exported. This is so that insert type cans succinctly
indicate what it thought should occur without failing (but still resulting in
code that causes another error).

Future work:

Extend annotate exports autofix to reserve import symbols to ensure
there are not collisions with other imports.

Reviewed By: nmote

Differential Revision: D16554662

fbshipit-source-id: ff161de6b2c2352d26ee502683f1d953992c8b1c