[enums] Custom error when incorrectly exporting enum like `export type enum`
Summary:
Some users may be used to using `export type` to export other Flow related things (like types): https://fb.workplace.com/groups/flow/permalink/3802665643115356/a

We don't use `export type` for enums as they are both types and values.

Add a custom error message to help users in the this case.

One thing I'm not sure about is if we should consume the `type` token and attempt to consume the enum, or not.

Reviewed By: mroch

Differential Revision: D21965665

fbshipit-source-id: 812d0ce1ab72b3c70c588cd498bab7d2f95bf8f0