basic JSDoc parser
Summary:
`Jsdoc.of_comments` parses a JSDoc from the leading comments on an AST node. If there are multiple leading comments that can be parsed as a JSDoc, it chooses the last one in the list.

[This website](https://jsdoc.app/about-getting-started.html) is the closest thing I could find to a spec for JSDoc. Keeping in line with the spec, we'll only pay attention to block comments starting with `/**`; this lets users opt in/out of whether their comment is picked up as documentation.

Currently this only parses the first portion of the jsdoc, before the tags. Later in this stack I'll expand it to parse certain tags we'll want to get information from, like param tags on functions.

Reviewed By: nmote

Differential Revision: D20776520

fbshipit-source-id: 89af223f035ebbcf072b3f143669388aecfdf39e