Extract documentation from ASTs
Summary: Makes the code that services the `hover` requests also call into the code that services `definition` requests; it then gives the getdef_loc to `Find_documentation.of_getdef_loc` which searches the relevant AST and returns the documentation associated with that definition if it finds it. In this diff the searcher only looks at `VariableDeclaration` nodes; future diffs will expand on this.

Reviewed By: mroch

Differential Revision: D21030003

fbshipit-source-id: c5faf2a6e5281abd984ca94a14c08b46ad741339