strongly type LSP exceptions
Summary:
`Lsp.Error.t` contains a currently untyped error code, which here I convert to an enum.

and instead of a collection of exceptions, use a single exception to make it easier to be sure we're handling them all. in the future I'd like to get rid of the exceptions in favor of a `result`, so that we enforce error handling.

Reviewed By: vrama628

Differential Revision: D21952309

fbshipit-source-id: c13d26a87ffa9aa1cf42afff8820d77e79301f8a