Add experimental CJS/ES interop mode for requiring a default ES export
Summary: Adds experimental CommonJS/ES module interop mode where `requires` of an ES module with a default export will be typed as that default export, instead of as an object with the default export behind the `default` key. This is automatically done if the ES module contains a default export (even if it contains other named exports e.g. types), and is gated behind the `experimental.module.automatic_require_default` flowconfig flag.

Reviewed By: panagosg7

Differential Revision: D21944567

fbshipit-source-id: 8b3f3a626b8ed41af0ea608e78d0cd73616b1fc6