Move expansion of attached coment locs into AST differ
Summary:
A while ago we made it so that attached comments are included when removing or replacing nodes in patches made from the AST differ. However in further testing this had some subtle bugs, as the expansion of locs to include attached comments was happening in the AST diff printer. There are some subtle cases where we actually need the expansion of locs to happen in the differ itself, such as when adding an `Insert` change after a `Replace` or `Delete` change in the `recurse_into_diff` utility.

Let's move the comment expansion down into the AST differ itself. To accomplish this we can move the utility that expands nodes' locs with attached comments in the AST differ itself, and then make sure that every created `Replace` or `Delete` change's loc is first run through this utility.

This is done by introducing the `replace` and `delete` utilities in the AST differ to properly expand the loc for replace and delete nodes to include the attached comments in the source text. Note that as before, an `Insert` change does not need its loc expanded as there is no node in the source text to refer to.

Reviewed By: pieterv

Differential Revision: D22004515

fbshipit-source-id: 8214e6ca8b85bf901b5aabf43fd02538c8d04b0a