[codemod] Work around babel parser issue when printing arrow functions with type params in a type cast
Summary:
The Babel parser currently has an issue where an arrow function with type params inside of a type cast is reported as a parse error. This is present in the current version of Babel (7.10), and I've opened an issue to track it: https://github.com/babel/babel/issues/11716.

```
// For example:
(<T>() => {}: any)
            ^ Parsing error: Unterminated JSX contents
```
Let's change our printer to work around this case so that printing and codemods will still be compatible with babel. Detecting this case in the `js_layout_generator` and then wrapping the arrow function in parens allows babel to parse the expression correctly.

Reviewed By: mroch

Differential Revision: D22026634

fbshipit-source-id: 8f7c599bc4ab80e6cb4c870f83d56597e7de549d