[easy] fix a comment
Summary:
this comment seems better positioned on `close_if_open` than on `close`, because that's where the `Lwt_unix.close` it's talking about actually is. the signature of `close ()` has nothing to do with it.

had to move `close_if_open` to a top-level function (which is fine anyway) to make ocamldoc happy.

Reviewed By: nmote

Differential Revision: D22052270

fbshipit-source-id: 393acb3eccde517b595b21b9c9641280e4efeb26