normalize paths in well_formed_exports.includes check
Summary:
The check for inclusion of a path in the `well_formed_exports.includes`
list would fail on windows.

This is because the path
```
well_formed_exports.includes=<PROJECT_ROOT>/a
```
would become
```
well_formed_exports.includes=/C:/actual/path/to/a
```
but files that were compared to it were of the form
```
\C:\path\to\file\b.js
```
With this change the latter paths are normalized to use `/` as path separator.

Reviewed By: nmote

Differential Revision: D22040870

fbshipit-source-id: 88cc02e32157f42dec758de4608232befd095fb0