[easy] rename EventLogger.log_if_initialized, which doesn't log
Summary:
`EventLogger.log_if_initalized f` doesn't log, it just calls `f` if the logger is initialized. so it would be better called `if_initalized f`, but we shouldn't reimplement programming language features... so it should be `if is_initialized () then f`. but that name is still not as clear as it should be: is being initialized sufficient to log?

so renamed it to `should_log`:

```
if should_log () then f
```

Reviewed By: nmote

Differential Revision: D22067335

fbshipit-source-id: c924f0a2a64e5920241b10ba41f8a625c30cbc82