fix shared memory exit codes from workers
Summary:
we currently have two sets of exit codes, `FlowExitStatus` and Hack's `Exit_status`.

in the worker code we inherited from Hack, we use `Exit_status`, but when we catch the exit in the server and/or monitor, we use `FlowExitStatus`. some of the codes are the same, like `Out_of_shared_memory` is 15 in both. But `Exit_status.Hash_table_full` (16) is `FlowExitStatus.Flowconfig_changed`.

if a worker dies, I rethrow it in the server (workerController), which should cause the server to die with the same code. due to this problem, we didn't report the correct exit reason in the server.

Reviewed By: samwgoldman

Differential Revision: D21998041

fbshipit-source-id: d32f222a6fc2bbb33928557db3528629f400c210