Make incremental_haste_excludes test test what it is supposed to
Summary:
This wasn't testing what the author thought it was testing. The file that was supposed to be excluded was actually included (see the changes to the .flowconfig). I noticed this while working on D22059258.

It appears that it has been broken since it was first committed in D4488705.

I needed to make some adjustments so that it actually tests what is intended. When `A.js` is renamed to `B.js`, `A` is no longer going to be found regardless of whether `B.js` is excluded or not. So, `index.js` has to be changed to require `B.js` as part of that change.

Reviewed By: samwgoldman

Differential Revision: D22059255

fbshipit-source-id: c9c4f59cd49c956a72aed8d73592f4044afb69f3