push watchman exceptions up the stack
Summary:
in order to recover from watchman issues, we need to stop exiting when they happen. we also need to make sure we catch all the various exceptions that can be raised. we can accomplish both by having the `Watchman` module use `result` more and exceptions less.

this diff is just some incremental refactoring I did to move exceptions closer to the surface.

Reviewed By: vrama628

Differential Revision: D22116709

fbshipit-source-id: fb540869a6753d1e1a9f2f159c49b35cf13b7cb8