Fix ast_diff_printer bug where separators in Insert changes are sometimes lost
Summary:
We currently have an issue with the `ast_diff_printer` where adjacent statements are sometimes printed on the same line. We can trace many of these occurrences back to an issue with the `flow_ast_differ` itself.

In the `standard_list_diff` algorithm, like-indexed `Inserts` and `Deletes` are converted to `Replaces`. However an `Insert` may contain a list of items, so if more than one item is contained within the `Insert` it will be split up into a `Replace` with the first item, followed by an `Insert` with the rest of the items.

The issue arises because `Inserts` contain an optional separator which is printed between adjacent elements in the `ast_diff_printer`. Since the first item of the `Insert` was broken out into a `Replace`, there is no separator printed between the `Replace` and the `Insert`, leading to the very first item in the original, unbroken `Insert` being printed on the same line as the second item.

We can get around this by adding a field to `Inserts` that tracks whether a leading separator should be printed. We then set this field to true when an `Insert` is split into a `Replace` followed by an `Insert`. The `ast_diff_printer` can then properly prepend a separator when printing `Inserts` if the leading separator field is true.

To make the use of `Insert` more clear, it is modified to be a record type and a helper function to construct it is introduced in the differ.

Reviewed By: pieterv

Differential Revision: D22142952

fbshipit-source-id: a0a4571e288d6bf2442db44b4830db13b515cf62