[PR] [DOCS] Remove Nuclide from docs for Atom
Summary:
According to official web site https://nuclide.io/ Nuclide is not supported anymore. After upgrading from old version of Atom it leads to an error. It is also not possible to install it anymore:

  Request for package information failed: Not Found

This will preven confusion and will save time for developers who still try to follow docs and use Atom with Nuclide for flow types.
Pull Request resolved: https://github.com/facebook/flow/pull/8420

Reviewed By: gkz

Differential Revision: D22303326

Pulled By: mroch

fbshipit-source-id: 5bb334a188d631b4a73d969dd412dbfa709926f6