[enums] Update type of `cast` and `isValid` to take in optional representation type
Summary:
Currently, the Flow Enum methods `isValid` and `cast` are typed to take in an argument of the representation type (e.g. for a string based enum, which are the default, this means the input has to currently be a string).

Observing some real-life use cases, I would like to expand the argument type to `?TRepresentationType`, to allow people to directly pass in things like optional properties of that type.

See also update to runtime: D22131221

Reviewed By: panagosg7

Differential Revision: D22138552

fbshipit-source-id: a1703ef9ebba9a8bf5831c02eb838ccb37e996f6