Fix: Include constant parts of EvalT destructors in SigHash
Summary:
This diff fixes the cases described in D22335901, by improving on
the hashing routine over `EvalT` types.

There were three cases in the `destructor` data type where important
fields were ignored in hashing: PropertyType, SpreadType and RestType. For
example, in SpreadType we did not account for the exactness of the new
type.

We also skipped over the top-level constructor of the EvalT which meant that
we could replace it in a code edit, while keeping the same payload, and
get the same hash.

This diff fixes all these cases.

Reviewed By: samwgoldman, jbrown215

Differential Revision: D22352911

fbshipit-source-id: 212b4ada545ca7b7044aefb1d006cb95b589a0d0