Remove unused, not-useful hh_mem_status API
Summary:
This API encodes an implementation detail which is that "absent" and "deleted"
hash table entries can be distinguished. The former describes entries which have
never been written to the hash table, while the latter describes entrie which
have been written, but were subsequently deleted.

The fact that deleted entries are encoded as a hash table slot with a non-zero
hash and a zero-valued addr is an implementation detail, and should not be
relied upon.

Happily, there is no reason to distinguish between these cases in practice. The
API was only used from tests to make unnecessary assertions. It is enough to
observe that hh_mem fails in either case.

Reviewed By: mroch

Differential Revision: D22368553

fbshipit-source-id: 408b57e47a361a761a5b0d5bb17ec5cf8e2e4be8