Preserve whitespace when printing statement lists in Insert nodes in ast_diff_printer
Summary:
Currently, the printing of toplevel statement lists by the `ast_diff_printer` does not preserve whitespace in the original statement list, instead printing the statements one after the other with no blank lines between them. This is often due to the fact that the statements are contained in a list within an `Insert` node, so they are all printed separately using `Js_layout_generator.statement`, which does not attempt to preserve existing whitespace. Instead we want to be using `Js_layout_generator.statement_list` when we detect that an `Insert` node contains a list of statements, as `Js_layout_generator.statement_list` will preserve whitespace in the original document.

There also exists one more case we must handle, which is when an `Insert` node is broken up into a `Replace` node with the first statement followed by an `Insert` node containing the rest of the statements. We can detect this case by checking when the `Replace` and `Insert` only contain statements, and if the location of the `Insert` is exactly the end loc of the `Replace`. In this case we should defer to `Js_layout_generator.statement_list` to preserve whitespace between statements (making this a single change instead of two, which is still correct since the insert is happening at the end of the replace).

Reviewed By: pieterv

Differential Revision: D22438571

fbshipit-source-id: 9248c4217ba8d930842dea5497c2db507f2c89bc