Import star invalid use rule errors on rest and computed properties in destructuring
Summary:
We have an optional lint rule which errors when a namespace import is used outside a static member access or destructuring, `import-star-invalid-use`.

This rule is meant to ensure that namespace imports are always treated as immutable collections of a module's named exports, where the namespace import can only be used via one of those named exports. We want to avoid any cases where exports are dynamically accessed from the namespace import object.

We allow accessing a namespace import's individual exports via a destructuring. However the current implementation is letting two rare cases of dynamism through - computed keys and rest elements in the object pattern. We should disallow computed keys in object patterns just like we disallow dynamic member accesses on the namespace object (but allow identifier and literal keys). We should also disallow having the rest element in the object pattern for namespace import destructuring, as we would then have to track the new rest object to make sure it is being used correctly as the same restrictions on the namespace import would need to apply.

Reviewed By: pieterv

Differential Revision: D22445968

fbshipit-source-id: f34601fcab8fa6e0e11371bab6ab66064ae4948b