Make get-def Type requests jump through AnnotT
Summary:
Get-def on an imported symbol should take you to the place in the source file where the symbol was defined. But in some cases, it would take you to the `module.exports = ...` line in the source file, which is less useful. This is particularly important for JSDoc integration, because finding JSDocs relies on get-def taking us to the definition of a symbol to look for a JSDoc there.

It looks like types-first is what caused this, and our get-def tests didn't pick up on it because they have types-first turned off. I added new get-def tests with types-first turned on and I observed that before this fix, it would jump to the `module.exports = ...` line rather than the actual definition.

Jumping through a `require` is handled by a `Get_def_request.Type` request. With types-first turned on, the type given to the `Type` request is wrapped in an `AnnotT` whose `def_loc` is on the `module.exports` line. This diff changes the way a `Type` request is serviced, so that it jumps through `AnnotT`s (in addition to jumping to `OpenT`s' lower bounds as before).

I think it's worth looking into how much other IDE behavior has changed because of types-first, and updating our IDE test suites to types-first since that's what most of our users are on.

Reviewed By: panagosg7

Differential Revision: D22464315

fbshipit-source-id: 1144f7128d53b815b1c4e93f72c1e658dd4da5b5