Expand project root token before creating Options.t
Summary:
Before this diff, the Options.t record stored configuration containing the
`<PROJECT_ROOT>` token which can be used in .flowconfig files. It was necessary
to replace that string with the project root before using it.

This diff performs that string replacement earlier, so that Options.t will not
contain the `<PROJECT_ROOT>` token.

Reviewed By: panagosg7

Differential Revision: D22517166

fbshipit-source-id: a2be5c9020a81f64cbd48fb85e347174010e18bd