Make --types-first cmdline option override well_formed_exports config
Summary:
Per D20454141 (https://github.com/facebook/flow/commit/a27ae2c261d847ddcfa8b050788cb24f41598283), types-first implies well-formed-exports. Indeed, it's an error to
set types_first=true in the .flowconfig if well_formed_exports=false.

Before this diff, passing the --types-first cmdline option would error if the
flowconfig well_formed_exports setting was explicitly set to false. This diff
changes the behavior to override the well_formed_exports setting in the
flowconfig instead.

We also need to deal with the case where `well_formed_exports.includes` config
is set. As described in D22520163, types-first requires well-formed exports
everywhere, so the override behavior also overrides the includes list.

Lastly, I combined two fields in the Options.t record to a single field. This
simplifies the code a bit.

This is edge case behavior, so the specific choice is not very important. I
prefer this behavior because (1) it's reasonable to expect the cmdline arguments
to override settings in the flowconfig and (2) if the user is passing
--types-first, they must want the behavior implied by well_formed_exports=true,
so let's give it to them automatically.

Reviewed By: panagosg7

Differential Revision: D22520203

fbshipit-source-id: bd904027313a80fd7f393dec3f2e5b0197151f1b