Lift tolerable_errors from File_sig.t
Summary:
Currently we store the list of tolerable errors in the File_sig.t record, but
they don't really need to live there. The tolerable errors are extracted from
the record before the file sig is saved into shared memory and are not needed
afterwards.

Instead, this diff changes the File_sig module to return the File_sig.t record
along with any tolerable errors on the side. These errors are combined with
signature errors to form the set of errors which should be printed when the
well-formed exports config is enabled.

Reviewed By: mroch

Differential Revision: D22521938

fbshipit-source-id: 44b6d6a13202d5f8421e3f7c5f5e1b2beba8dded