Add new type-sig heap
Summary:
For the first phase of types-first 2.0, we will introduce a new mode that uses
type signatures instead of SigAST and SigFileSig. Like types-first 1.0, we will
store the signatures at the granularity of files, so we can continue to use the
existing sharedmem implementation.

This diff only defines the heaps and provides a mechanism to write to them, but
nothing will write to them yet. Later diffs will continue to flesh this out.

Reviewed By: mroch

Differential Revision: D22563931

fbshipit-source-id: 6d7545d1a7b242fe7ce9beac670d8e2cfe9f4b55