delete broken dynamic-export lint rule
Summary: This lint rule was added to detect when an any type was being exported in a file export, but was not used in practice because it triggered so often. It is, however, broken in types first and without any uses is not worth maintaining. Hence we delete it instead.

Reviewed By: samwgoldman

Differential Revision: D22672653

fbshipit-source-id: 7044d653caffcbe2977d6c6e419cd686485534a2