Improve error message for non-strict equality check
Summary:
We are very strict in what we accept to be compared with `==` and `!=` because of the implicit coercions JS does for them. We allow almost anything to be compared with strict equality (`===` & `!==`) however.

The previous message (shared with errors for strict equality and also the `<`, `>` operators) of simply "Cannot compare X and Y" was not helping guide the user to fix their issue (either, change the arguments, or use strict equality instead).

Open to suggestions on wording!

Reviewed By: panagosg7

Differential Revision: D22583739

fbshipit-source-id: e0ba3928594e24612d4afc00825ce62b17116364