Simplify chaining logic in get-def
Summary:
This diff simplifies the logic of get-def by
- deleting `special_caser`, one of the two AST searchers
- Making decisions about "chaining" happen in a top-down way, instead of having helpers make decisions about whether the results they produce are final or need to be chained.

The `special_caser` searcher I've deleted used to have two responsibilities, each of which were non-essential: 1) to notice when the result of get-def comes from an import statement, so that we can `Chain` another request to follow the import. Since the `get_def_process_location` searcher basically also does this, we can just use it instead. 2) to reposition things in minor ways (see lines 99-130 of the old getDef_js.ml). This was only added to keep consistent with the old behavior from when I refactored get-def off of the type inference hooks, but I don't think the old behavior is particularly necessary to conserve (lmk if you disagree).

We used to let helpers return `Chain <request>` to indicate that the result they found isn't final and another request needs to be done. This was unnecessary; now we just always call `get_def_process_location` on the result of the helper and it decides whether further requests need to be done.

Feel free to nitpick

Reviewed By: gkz

Differential Revision: D22194407

fbshipit-source-id: e039a3f4b9fe5f8bf71a5fd032dff1dc74a0c167