Allow JSDocs to have no description
Summary:
Sometimes a JSDoc doesn't have a description; e.g.:
```
/**
 * param foo bar
 */
```
This diff changes the `description` in a JSDoc to be `string option`, to more accurately reflect the fact that a JSDoc can have no description. Previously when there was a JSDoc with no description we would tell hover that the description was the empty string, and VSCode happened to be smart enough to not render that weirdly. A result of this diff is that now when there is no description, we (correctly) don't tell hover that there is a description.

Reviewed By: gkz

Differential Revision: D22667388

fbshipit-source-id: 041e3a8b1f1ea5183793b1333565db11df34ffbd