Find JSDocs in object literals
Summary:
find documentation on members of object literals.

I noticed while working on this that get-def doesn't work very well on members of imported object literals, because types-first forces object literals to be type-annotated, which throws off get-def by pointing to the member in the object type instead of the member in the actual object expression. This should probably be fixed, but it seems like it might be tough.

Another (much smaller) confusing bit of behavior is where get-def of a getter property without a return type annotation jumps to; it doesn't seem to have a meaningful behavior.

Reviewed By: gkz

Differential Revision: D22652715

fbshipit-source-id: 4e949f1d753cee9178a5accbd375e70764675c7a