make get-def of destructured imports go to declaration of import
Summary:
```
// exports.js
let foo = ...; // where this diff makes get-def go
module.exports = {foo} // where get-def used to go

// imports.js
const {foo} = require('./exports');
foo // <-- get-def request here
```

A common pattern for CommonJS exports is to declare a bunch of stuff in a file then export them in an object literal, and import them using destructuring. When a get-def request is issued on something imported this way, we used to jump to the identifier in the `module.exports = {...}` line, but we've gotten feedback that it'd be more useful to jump up to the declaration in the imported file. This diff fixes that.

Reviewed By: gkz

Differential Revision: D22577950

fbshipit-source-id: d31d394824fddb03f5d8bb6313cb26c0a9913811