relax "bad exports" errors for ES modules
Summary:
BadExportPosition and BadExportContext are errors accumulated when creating
file-sig and refer to bad uses of `module` and `module.exports` that would expose
them to aliasing (see tests/poorly_formed_exports). I believe they should only
target CommonJS exports, since for ES we do not use `module.exports`
to compute the exported signature.

This diff does this. The filtering of the overly strict errors happens after the traversal
of the AST, since we need to be certain of the kind of the module.

Fixes https://github.com/facebook/flow/issues/8442

Reviewed By: samwgoldman

Differential Revision: D22675807

fbshipit-source-id: 18ccc0ce0b04dfb891ee5a4d8b356b7d8e7d54fa